+++ lib/l10n/translations/gsw.dart
@@ -1785,6 +1785,7 @@ const _dutchSourceGsw = {
 };
 
 const _dutchSourceAddGsw = <String, String>{
+  'Media verwijderd om privacyredenen': 'Medie us Dateschutzgründ entfernt',
   'Afbeeldingen door AI laten taggen?': 'Söll d\'KI die Bilder tagge?',
   'afbeeldingen gaan naar': 'Bilder gönd a',
   'een model op dit apparaat': 'es Modäll uf däm Grät',