+++ lib/l10n/translations/tlh.dart
@@ -1756,6 +1756,7 @@ const _dutchSourceTlh = {
 };
 
 const _dutchSourceAddTlh = <String, String>{
+  'Media verwijderd om privacyredenen': 'So\'wI\' meq luteqta\' media',
   'Afbeeldingen door AI laten taggen?': 'nagh HaSta pong\'a\' AI?',
   'afbeeldingen gaan naar': 'HaSta lupoQ',
   'een model op dit apparaat': 'jan vam model',